Add YOLO2COCO and COCO2LS conversion scripts
This commit adds the YOLO2COCO and COCO2LS conversion scripts to the project. These scripts allow for the conversion of datasets from YOLOV5 format to COCO format and vice versa. The scripts can be found in the `bin` directory. Additionally, the commit includes the necessary changes to the `__init__.py` and `base.py` files in the `phenocv/convert` directory to import and use these conversion scripts. Finally, the commit also includes a new function `match_yolo_annotations` in the `phenocv/utils.py` file that matches YOLO annotations with corresponding images and removes any unmatched files.